requirements.txt
appdirs==1.4.4
asgiref==3.4.1
boto3==1.17.57
botocore==1.20.57
Django==3.2.5
django-filter==2.4.0
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.1.2
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.3.0
